fix: выпускной проект часть 3 - корректировка и улучшение

+++ ewm-service/src/main/java/ru/practicum/users/service/UserServiceImpl.java
@@ -30,14 +30,14 @@ public class UserServiceImpl implements UserService {
     @Override
     public UserDto createUser(NewUserRequest userDto) {
         User user = userRepository.save(toUser(userDto));
-        log.info("Создан пользователь {}", user);
+        log.info("Соз0дан пользователь {}", user);
         return toUserDto(user);
     }
 
     @Transactional(readOnly = true)
     @Override
-    public List<UserDto> getUsers(List<Long> ids, Integer from, Integer size) {
-        log.info("Получены пользователи со следующими ids: {}, from {}, size {}", ids, from, size);
+    public List<UserDto> getAllUsers(List<Long> ids, Integer from, Integer size) {
+        log.info("Получен список пользователей с ids: {}, from {}, size {}", ids, from, size);
         if (ids.isEmpty()) {
             return userRepository.findAll(new Pagination(from, size, Sort.unsorted())).stream()
                     .map(UserMapper::toUserDto)
@@ -51,9 +51,12 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUserById(Long userId) {
-        userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundException("Пользователь с id=" + userId + " не найден"));
-        log.info("Удаленр пользоватлеь с id= {}", userId);
+        if (!userRepository.existsById(userId)) {
+            throw new NotFoundException("Пользователь с id = " + userId + " не был найден");
+        }
+
         userRepository.deleteById(userId);
+        log.info("Пользователь с id = {} успешно удален.", userId);
     }
+
 }
